Add files via upload
I followed a tutorial on YouTube and typed all of the included code up to the part about the dog. The formatting in the video made some of the longer lines of code difficult to see and I hit a roadblock. At any rate, I'm developing a better understanding of Python and PyCharm.