My first go at React
I spent an evening following along with a TeamTreeHouse course on coding a scoreboard. I'm still fairly new to the world of computer programming, but the two files (app.css and app.jsx) are the result of my time. I know they won't load or run without the other files, but I just wanted to preserve the code that I typed. I made a lot of mistakes along the way, and I spent about half of my time just finding and fixing said errors. I love this process so much. The feeling that I get when I hit run, compile, or refresh and it works is unbeatable!!!